Split HDF5 wrapper into individual files
According to HDF5 concept
