+++ src/merger_tree_reader.cpp
@@ -41,7 +41,7 @@
 #include "subhalo.h"
 #include "timer.h"
 #include "utils.h"
-#include "hdf5/reader.h"
+#include "hdf5/io/reader.h"
 
 
 namespace shark {
